// src/frontend/src/app/api/reports/swagger.ts
export enum WiState {
    Design = 0,
    Closed = 1,
    Ready = 2,
}

export enum AutomationStatus {
    Manual = 0,
    Automated = 1,
}
// ...
export class AddOrUpdateTestCaseRequest implements IAddOrUpdateTestCaseRequest {
    priority!: number;
    state!: WiState;
    assignedTo!: number;
    title?: string | undefined;
    description?: string | undefined;
    automationStatus!: AutomationStatus;
    automationTestName?: string | undefined;
    automationTestStorage?: string | undefined;
    automationTestType?: string | undefined;
    precondition?: string | undefined;
    postcondition?: string | undefined;
    steps?: AddOrUpdateTestStep[] | undefined;

    constructor(data?: IAddOrUpdateTestCaseRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.priority = _data["priority"];
            this.state = _data["state"];
            this.assignedTo = _data["assignedTo"];
            this.title = _data["title"];
            this.description = _data["description"];
            this.automationStatus = _data["automationStatus"];
            this.automationTestName = _data["automationTestName"];
            this.automationTestStorage = _data["automationTestStorage"];
            this.automationTestType = _data["automationTestType"];
            this.precondition = _data["precondition"];
            this.postcondition = _data["postcondition"];
            if (Array.isArray(_data["steps"])) {
                this.steps = [] as any;
                for (let item of _data["steps"])
                    this.steps!.push(AddOrUpdateTestStep.fromJS(item));
            }
        }
    }

    static fromJS(data: any): AddOrUpdateTestCaseRequest {
        data = typeof data === 'object' ? data : {};
        let result = new AddOrUpdateTestCaseRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["priority"] = this.priority;
        data["state"] = this.state;
        data["assignedTo"] = this.assignedTo;
        data["title"] = this.title;
        data["description"] = this.description;
        data["automationStatus"] = this.automationStatus;
        data["automationTestName"] = this.automationTestName;
        data["automationTestStorage"] = this.automationTestStorage;
        data["automationTestType"] = this.automationTestType;
        data["precondition"] = this.precondition;
        data["postcondition"] = this.postcondition;
        if (Array.isArray(this.steps)) {
            data["steps"] = [];
            for (let item of this.steps)
                data["steps"].push(item.toJSON());
        }
        return data;
    }
}

export interface IAddOrUpdateTestCaseRequest {
    priority: number;
    state: WiState;
    assignedTo: number;
    title?: string | undefined;
    description?: string | undefined;
    automationStatus: AutomationStatus;
    automationTestName?: string | undefined;
    automationTestStorage?: string | undefined;
    automationTestType?: string | undefined;
    precondition?: string | undefined;
    postcondition?: string | undefined;
    steps?: AddOrUpdateTestStep[] | undefined;
}

export class AddOrUpdateTestStep implements IAddOrUpdateTestStep {
    id?: number | undefined;
    order!: number;
    action?: string | undefined;
    result?: string | undefined;

    constructor(data?: IAddOrUpdateTestStep) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.order = _data["order"];
            this.action = _data["action"];
            this.result = _data["result"];
        }
    }

    static fromJS(data: any): AddOrUpdateTestStep {
        data = typeof data === 'object' ? data : {};
        let result = new AddOrUpdateTestStep();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["order"] = this.order;
        data["action"] = this.action;
        data["result"] = this.result;
        return data;
    }
}

export interface IAddOrUpdateTestStep {
    id?: number | undefined;
    order: number;
    action?: string | undefined;
    result?: string | undefined;
}